% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_mmm.R
\name{average_hessians}
\alias{average_hessians}
\title{Average the Hessians}
\usage{
average_hessians(derivatives, pairs, data, id, A)
}
\arguments{
\item{derivatives}{the derivatives returned by get_mmm_derivatives().}

\item{pairs}{A character matrix with pairs returned by the make_pairs function.}

\item{data}{The data.frame with original data in long format.}

\item{id}{A character value with the subject identifier in the orginal data.}

\item{A}{the A helper matrix returned by the create_A_matrix function.}
}
\value{
a Sparse matrix of the MMM Hessian
}
\description{
...
}
