% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_mmm.R
\name{binary.normal}
\alias{binary.normal}
\title{Link function for the combination of a binary and continuous marker}
\usage{
binary.normal(indicator = stop("'indicator must be specified'"))
}
\description{
Link function for the combination of a binary and continuous marker
}
