% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_mmm.R
\name{create_A_matrix}
\alias{create_A_matrix}
\title{Create the helper 'A'  matrix used to average the parameter estimates.}
\usage{
create_A_matrix(parameter_labels)
}
\arguments{
\item{parameter_labels}{A data frame with parameter labels returned by the stack_parameters() function.}
}
\value{
A p*q matrix with ones if a parameter was estimated and zeros if it was not estimated in the pairwise step
}
\description{
Take parameter labels from MMM object and return a p*q matrix where p is unique(parameters) and q is the total number of parameters (including duplicates).
}
