% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance_mmm.R
\name{get_auc}
\alias{get_auc}
\title{Calculate the Area under the Receiver Operating Characteristics Curve}
\usage{
get_auc(
  predictions,
  prediction_landmark,
  prediction_horizon,
  id,
  failure_time,
  failure
)
}
\arguments{
\item{predictions}{predictions data from the mmm_predictions function.}

\item{prediction_landmark}{a numeric vector with the landmark times to select predictions.}

\item{prediction_horizon}{the prediction horizon \emph{at} which to calculate the ROC-AUC.}

\item{id}{a character value for the name of the vector of subject identifiers in the predictions data.}

\item{failure_time}{a character value for the name of the vector with failure times in the predictions data.}

\item{failure}{a character value for th name of the vector with failure status.}
}
\description{
Use the timeROC package to calculate the ROC-AUC based on predictions
from the mmm_predictions function.
}
\details{
Uses the timeROC package to determine the area under the receiver operating characteristics curve.
}
