% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_mmm.R
\name{get_covariance}
\alias{get_covariance}
\title{Calculate the covariance matrix for the MMM and return it and the robust standard error}
\usage{
get_covariance(hessian, cpgradient, parameters, A, data, id)
}
\arguments{
\item{hessian}{The sparse Hessian matrix for the MMM returned by the average_hessians function().}

\item{cpgradient}{The cross-product and average of the subject level gradients returned by the get_crossproducts() function.}

\item{parameters}{The parameter estimates returned by the stack_parameters() function.}

\item{A}{the A helper matrix returned by the create_A_matrix function().}

\item{data}{The data.frame with original data in long format.}

\item{id}{A character value with the subject identifier in the orginal data.}
}
\value{
A data.frame with parameter estimates and standard errors for the multivariate mixed model.
}
\description{
Use a sandwich estimator to return the standard errors for the full multivariate mixed model.
}
