% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_mmm.R
\name{get_crossproducts}
\alias{get_crossproducts}
\title{Get cross products of the gradients}
\usage{
get_crossproducts(pairs, parameter_labels, gradients, data, id)
}
\arguments{
\item{pairs}{A character matrix with pairs returned by the make_pairs function.}

\item{parameter_labels}{The parameter labels returned by the stack_parameters() function.}

\item{gradients}{A data frame with gradients returned by the stack_gradients() function.}

\item{data}{The data.frame with original data in long format.}

\item{id}{A character value with the subject identifier in the orginal data.}
}
\value{
A matrix with the average gradient for all parameters over all subjects.
}
\description{
Get the cross product of the subject specific gradient vector, sum and take the average
}
