% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_mmm.R
\name{get_likelihood_fail}
\alias{get_likelihood_fail}
\title{Get Likelihood Fail}
\usage{
get_likelihood_fail(
  data,
  outcomes,
  fixed_formula_fail = NULL,
  random_formula_fail = NULL,
  time,
  failure_time,
  parameters_fail,
  random_effects_fail,
  id,
  landmark,
  horizon,
  interval,
  ...
)
}
\arguments{
\item{data}{data.frame with original data in long format}

\item{outcomes}{a character vector with the labels for the longitudinal outcomes.}

\item{fixed_formula_fail}{a character string with lme4 style formula for the fixed effects}

\item{random_formula_fail}{a character string with lme4 style formula for the random effects}

\item{time}{a character string with the label for the follow-up time variable}

\item{failure_time}{a character string with the label for the failure time variable}

\item{parameters_fail}{a data.frame with the parameter labels and estimates returned by mmm_model()}

\item{random_effects_fail}{The random effects sampled from a multivariate normal distribution (see MASS::mvrnorm)}

\item{id}{a character string with subject identifier}

\item{landmark}{the landmark time until which data is available.}

\item{horizon}{the prediction horizon}

\item{interval}{the intervals at which predictions are to be made}
}
\value{
A list with the likelihood contributions for each subject (rows) at every time point (cols).
}
\description{
Get the parameter estimates from a MMM and combine with data to return contributions to the likelihood at observed time points for non failures
}
