% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_mmm.R
\name{get_likelihood_nofail}
\alias{get_likelihood_nofail}
\title{Get Likelihood No Fail}
\usage{
get_likelihood_nofail(
  data,
  outcomes,
  fixed_formula_nofail = NULL,
  random_formula_nofail = NULL,
  time,
  parameters_nofail,
  random_effects_nofail,
  id,
  horizon,
  ...
)
}
\arguments{
\item{data}{data.frame with original data in long format}

\item{outcomes}{a character vector with the labels for the longitudinal outcomes.}

\item{fixed_formula_nofail}{a character string with lme4 style formula for the fixed effects}

\item{random_formula_nofail}{a character string with lme4 style formula for the random effects}

\item{time}{a character string with the label for the follow-up time variable}

\item{parameters_nofail}{a data.frame with the parameter labels and estimates returned by mmm_model()}

\item{random_effects_nofail}{The random effects sampled from a multivariate normal distribution (see MASS::mvrnorm)}

\item{id}{a character string with subject identifier}

\item{horizon}{the prediction horizon}
}
\value{
A list with the likelihood contributions for each subject (rows) at every time point (cols).
}
\description{
Get the parameter estimates from a MMM and combine with data to return contributions to the likelihood at observed time points for non failures
}
