% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_mmm.R
\name{get_likelihood_profiles_fail}
\alias{get_likelihood_profiles_fail}
\title{Get likelihood profiles for failure}
\usage{
get_likelihood_profiles_fail(likelihood_fail, landmark, horizon, interval)
}
\arguments{
\item{likelihood_fail}{a list of likelihoods for each subject at the observed time points}

\item{landmark}{the prediction landmark time.}

\item{horizon}{the prediction horizon.}

\item{interval}{the intervals relative to the horizon (e.g. 1/12 for monthly intervals and horizon is in years)}
}
\value{
A list of length (horizon - landmark)/interval of lists with nrow(unique(id)) with likelihood profiles
}
\description{
Generate a step function of the likelihood at interval times t.
}
