% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_mmm.R
\name{get_mmm_derivatives}
\alias{get_mmm_derivatives}
\title{Determine the subject level derivatives.}
\usage{
get_mmm_derivatives(
  stacked_data,
  id,
  fixed,
  random,
  pairs,
  model_families,
  start_values,
  nAGQ = 11
)
}
\arguments{
\item{stacked_data}{A list of data.frames returned by the stack_data() function}

\item{id}{The name of the column with subject ids in stacked_data}

\item{fixed}{A formula (as character string) for the fixed part of the mixed model. Passed to as.formula().}

\item{random}{A formula (as character string) for the random part of the mixed model. Passed to as.formula().}

\item{pairs}{A character matrix with pairs returned by the make_pairs function.}

\item{model_families}{A list with the model families returned by the test_input_datatypes() function.}

\item{start_values}{A list (length == length(stacked_data)) with start values returned by the get_start_values() function.}

\item{nAGQ}{The number of knots to use on the Adapative Gaussian Quadrature (passed to GLMMadaptive::mixed_model()).}
}
\value{
A list of length nrow(pairs) each with two elements:
\itemize{
\item{df_hessians}{a list with data.frames storing the subject level Hessians}
\item{df_gradients}{a list with data.frames storing the subject level gradients}
}
}
\description{
Run mixed models without iterating and return subject level derivatives
}
\details{
This function uses the start values determined in by the function get_start_values()
to determine the subject level contributions to the derivatives. In turn these are used
to average the parameters and standard errors for the full multivariate mixed model.
}
