% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_mmm.R
\name{get_mmm_start_values}
\alias{get_mmm_start_values}
\title{Fit pairwise mixed models to get start values for the multivariate mixed model}
\usage{
get_mmm_start_values(
  stacked_data,
  fixed,
  random,
  pairs,
  model_families,
  user_initial_values = NULL,
  nAGQ = 11,
  iter_EM = 30,
  iter_qN_outer = 10,
  tol1 = 0.001,
  tol2 = 1e-04,
  tol3 = 1e-08,
  ...
)
}
\arguments{
\item{stacked_data}{a list with the stacked data returned by the stack_data() function.}

\item{fixed}{character string with the fixed model part. This is passed to as.formula().}

\item{random}{character string with the random model part. This is passed to as.formula().}

\item{pairs}{a character matrix with the pairs returned by the make_pairs() function.}

\item{model_families}{a list with family names and indicators returned by the test_input_datatypes() function.}

\item{...}{arguments passed to GLMMadaptive::mixed_model()}
}
\value{
a list with starting values to determine the subject level contributions to the derivates.
}
\description{
Takes the stacked data and the user defined formulas for the fixed and random effects. Leverages the future.apply package to parallelize along pairs.
Using a different number of workers than pairs may result in problems during the optimization process.
}
