% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_mmm.R
\name{get_model_output}
\alias{get_model_output}
\title{Create the full mixed model output}
\usage{
get_model_output(estimates)
}
\arguments{
\item{estimates}{A data.frame with estimates returned by the get_covariance() function.}
}
\value{
A list of 3 elements including:
\describe{
\item{estimates}{A data.frame with the parameter estimates and standard errors}
\item{vcov}{The Variance-Covariance matrix for the random effects}
\item{corr}{A correlation matrix for the random effects}
}
}
\description{
Take the random effect estimates and return vcov and correlation matrices
}
