% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_mmm.R
\name{get_outcome_type}
\alias{get_outcome_type}
\title{Return a data.frame with the outcome label and type}
\usage{
get_outcome_type(data, outcomes)
}
\arguments{
\item{data}{original data}

\item{outcomes}{a character vector with the names of the longitudinal outcomes}
}
\description{
Return a data.frame with the outcome label and type
}
