% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_mmm.R
\name{get_posteriors}
\alias{get_posteriors}
\title{Apply Bayes Rule to get posterior predictions}
\usage{
get_posteriors(f_fail, f_nofail, prior)
}
\arguments{
\item{f_fail}{density for the failures at landmarks upto horizon.}

\item{f_nofail}{density for the non-failures up to horizon.}

\item{prior}{the priors returned by get_priors.}
}
\value{
: a list with posterior predictions at every landmark
}
\description{
Use the posteriors for failures and non-failure to get predictions with Bayes Rule
}
