% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_mmm.R
\name{get_predictions_fail}
\alias{get_predictions_fail}
\title{Get Predictions for Failure}
\usage{
get_predictions_fail(likelihood_profiles_fail, prior)
}
\arguments{
\item{likelihood_profiles_fail}{a list with the likelihood profiles for the failure model for each subject}

\item{prior}{a data frame with priors returned by get_priors()}
}
\value{
a list of dataframes with the predictions for each subject up to the horizon.
}
\description{
Take a cumsum of the posteriors and return the cumulative incidence of the event between interval and horizon
Implements equation 7.2 (p155) from Steffen's thesis.
}
