% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_mmm.R
\name{get_predictions_nofail}
\alias{get_predictions_nofail}
\title{Get predictions for non failure}
\usage{
get_predictions_nofail(likelihood_profiles_nofail)
}
\arguments{
\item{likelihood_profiles_fail}{a list with the likelihood profiles for the failure model for each subject}
}
\value{
a list of dataframes with the predictions for each subject at every interval up to the horizon
}
\description{
Take a cumsum of the posteriors and return the cumulative incidence of the event between interval and horizon
}
