% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_mmm.R
\name{get_priors}
\alias{get_priors}
\title{Get priors}
\usage{
get_priors(data, time_failure, failure, horizon, interval, dist = "weibull")
}
\arguments{
\item{data}{a data.frame with the data for which to predict in the long format (input should be the same as likelihood_profiles).}

\item{time_failure}{a character with the failure time variable name.}

\item{failure}{a character with the failure indicator variable name.}

\item{horizon}{a 1L int with the prediction horizon.}

\item{interval}{the intervals relative to the horizon (e.g. 1/12 for monthly intervals and horizon is in years)}

\item{dist}{distribution parameter for the survreg()}
}
\value{
a priori survival and failure probabilities (1-survival) for every interval between min landmark and max horizon
}
\description{
Call flexsurvreg to fit an unconditional survival model and return obtain priors
}
\examples{
\dontrun{
# Return prior up to a horizon of 5 (years) at 3 month intervals.
get_priors(data=df,
  time_failure="ftime",
  failure="failure",
  horizon=5,
  interval=1/4
)
}
}
