% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_mmm.R
\name{get_random_start}
\alias{get_random_start}
\title{Create random start values}
\usage{
get_random_start(model_families, n_fixed, n_random)
}
\arguments{
\item{model_families}{a list with model inf returned by test_input_datatypes}

\item{n_fixed}{the number of fixed parameters}

\item{n_random}{the number of random parameters}
}
\value{
a list with model information including:
\itemize{
\item{The model family.}
\item{An indicator for the binary outomce in the binary.normal family.}
\item{Starting values for the model fitting stage.}
}
}
\description{
Take a model info object created by test_input_datatypes and add random start values
}
