% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_data.R
\name{make_pairs}
\alias{make_pairs}
\title{Make pairs}
\usage{
make_pairs(outcomes)
}
\arguments{
\item{outcomes}{a character vector with the names of longitudinal outcomes}
}
\value{
a matrix of length(outomes)*(length(outcomes)-1)/2 rows and 2 columns with all unique pairs.
}
\description{
Take a character with the names of the longitudinal outcome variables and make all unique pairs
}
