% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_mmm.R
\name{make_random_formula}
\alias{make_random_formula}
\title{Generate random effects formula for stacked data}
\usage{
make_random_formula(id, covars = NULL)
}
\arguments{
\item{id}{a character value with the name of the vector of subject identifiers.}

\item{covars}{a list of character vectors with the covariates names}
}
\value{
a list with the fixed formulas for the pairwise model fitting stage.
}
\description{
Takes a list of covariates for the random slopes.
Each element of the list are the random slopes for a pairwise model
}
\examples{
# Random intercept at subject level
random <- make_random_formula(id = "id")

# Add random slopes
\dontrun{
random <- make_random_formula(id = "id", covars = "time")
}
}
