% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance_mmm.R
\name{plot_calibration}
\alias{plot_calibration}
\title{Plot the calibration for MMM}
\usage{
plot_calibration(
  predictions,
  prediction_landmark,
  prediction_horizon,
  id,
  failure_time,
  failure
)
}
\arguments{
\item{predictions}{the predictions returned by the mmm_predictions function}

\item{prediction_landmark}{a numeric vector with the landmark times to select predictions}

\item{prediction_horizon}{the prediction horizon \emph{at} which to calculate the ROC-AUC.}

\item{id}{a character value for the name of the vector of subject identifiers in the predictions data}

\item{failure_time}{a character value for the name of the vector with failure times in the predictions data}

\item{failure}{a character value with the name of the vector of failure indicators.}
}
\value{
a list of grobs
}
\description{
Create a grobs with calibration plot for a given landmark and horizon.
}
