% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance_mmm.R
\name{plot_predictions}
\alias{plot_predictions}
\title{Plot predictions from the multivariate mixed model}
\usage{
plot_predictions(predictions, outcomes, id, subject, time)
}
\arguments{
\item{predictions}{the predictions generated from mmm_predictions()}

\item{outcomes}{a character vector with the names of the longitudinal outcomes}

\item{id}{a character value with the name of the subject identifier}

\item{subject}{a character value with id for subject of interest}

\item{time}{a scalar value for the time at which the prediction is plotted}
}
\value{
plots for survival conditional on the longitudinal biomarker trajectory.
}
\description{
For a given landmark time, plot the probability of failure conditional on survival
and longitudinal biomarker evolutions.
}
