% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_data.R
\name{stack_data}
\alias{stack_data}
\title{Return a list with stacked data sets for each of the outcome pairs}
\usage{
stack_data(data, id, pairs, covars = NULL)
}
\arguments{
\item{data}{A data.frame with the longitudinal data in long format.}

\item{id}{A character value with the variable name of the subject identifier.}

\item{pairs}{A character matrix with the pairs, returned from the make_pairs function}

\item{covars}{(optional) a character vector with the names of the covariate vectors.}
}
\value{
a list of data.frames of length nrow(pairs).
}
\description{
Takes a data.frame in with longitudinal data to be used in the multivariate mixed model function.
Uses the pairwise combination created with the make_pairs() function to unroll the outcome matrix into a vector.
The longitudinal predictor and time invariant covariates are processed accordingly.
}
\examples{
\dontrun{
df_stacked <- stack_data(
  data = df,
  id = "id",
  pairs = pairs,
  covars = c("time", "sex", "time_failure")
)
}
}
