% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_mmm.R
\name{stack_gradients}
\alias{stack_gradients}
\title{Stack gradients into a single data.frame}
\usage{
stack_gradients(derivatives, data, id, pairs)
}
\arguments{
\item{derivatives}{list with derivatives returned by the get_mmm_derivatives() function.}

\item{data}{The data.frame with original data in long format.}

\item{id}{A character value with the subject identifier in the orginal data.}

\item{pairs}{A character matrix with pairs returned by the make_pairs function.}

\item{parameter_labels}{The parameter labels returned by the stack_parameters() function.}
}
\value{
A data.frame with subject level gradients for every longitudinal outcome
}
\description{
Take the gradients returned by the get_mmm_derivatives function and stack these into a single data.frame for further processing
}
