% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_mmm.R
\name{stack_parameters}
\alias{stack_parameters}
\title{Take the parameter estimates, add labels and stack them into a data.frame}
\usage{
stack_parameters(start_values, pairs, model_families)
}
\arguments{
\item{start_values}{The list of start_values (i.e. parameter estimates) returned by the get_start_values() function.}

\item{pairs}{A character matrix with pairs returned by the make_pairs function.}

\item{model_families}{the model_families object returned by test_input_datatypes()}
}
\value{
A data.frame with the labelled parameter estimates.
}
\description{
Extract the parameter estimates from the start_values and label them back to the names of the original data.
}
