% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_data.R
\name{test_compare_stacked_to_original_data}
\alias{test_compare_stacked_to_original_data}
\title{Test to compare stacked outcome data to original outcome data}
\usage{
test_compare_stacked_to_original_data(data, stacked_data, pairs)
}
\arguments{
\item{data}{data.frame with original data in long format}

\item{stacked_data}{the list of data.frames returned by stack_data()}

\item{pairs}{A character matrix with the pairs, returned from the make_pairs function}
}
\value{
Prints a list with assertions. All should return TRUE. Otherwise something went wrong.
}
\description{
...
}
