% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_data.R
\name{test_input_datatypes}
\alias{test_input_datatypes}
\title{Test input data types and return model families and indicator for binary outcomes}
\usage{
test_input_datatypes(data, pairs)
}
\arguments{
\item{data}{A data.frame in long format}

\item{pairs}{A character matrix constructed by make_pairs function.}
}
\value{
A list of length nrow(pairs) that includes
\describe{
\item{families}{character vector with the custom model family.}
\item{indicator}{a character vector with the indicator for the binary outcome if family = binary.normal}
}
}
\description{
Takes a data.frame with longitudinal data in long format and checks if the longitudinal data are binary or numeric.
}
