% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_mmm.R
\name{mmm_predictions}
\alias{mmm_predictions}
\title{Predictions from the multivariate mixed model using a disciminant analysis framework.}
\usage{
mmm_predictions(
  data,
  outcomes,
  fixed_formula_nofail,
  random_formula_nofail,
  random_effects_nofail,
  parameters_nofail,
  fixed_formula_fail,
  random_formula_fail,
  random_effects_fail,
  parameters_fail,
  time,
  failure,
  failure_time,
  prior,
  id,
  landmark = 0,
  horizon,
  interval,
  ...
)
}
\arguments{
\item{data}{a data.frame with the data for which to predict in the long format (input should be the same as likelihood_profiles)}

\item{outcomes}{character vector with the outcomes}

\item{fixed_formula_nofail}{formula for the fixed effects of the nofail model}

\item{random_formula_nofail}{formula for the random effects of the nofail model,}

\item{random_effects_nofail}{random effect estimates for the no fail model (see MASS::mvnorm)}

\item{parameters_nofail}{parameter estimates from the non-failures model (see mmm_model)}

\item{fixed_formula_fail}{formula for the fixed effects of the fail model,}

\item{random_formula_fail}{formula for the random effects of the fail model,}

\item{random_effects_fail}{random effect estimates for the fail model (see MASS::mvnorm),}

\item{parameters_fail}{parameter estimates from the non-failures model (see mmm_model)}

\item{time}{a character string with the time variable name in data}

\item{failure}{a character string with the failure variable name in data}

\item{failure_time}{a character string with the failure time variable in data}

\item{prior}{the priors returned by get_priors().}

\item{id}{a character string with the subject id}

\item{landmark}{a numeric value indicating the prediction landmark until which data is available for prediction.}

\item{horizon}{the prediction horizon (maximum follow-up time)}

\item{interval}{the intervals relative to the horizon (e.g. 1/12 for monthly intervals and horizon is in years)}
}
\value{
a data.frame with the posterior predictions for each subject at every landmark up to the horizon
}
\description{
Calculate the posterior predictions from the likelihood profiles and priors. Implements equation 7.1 (p154) from Steffen's thesis.
}
\examples{
\dontrun{
# First retrieve random effects samples from a fitted model
mu <- setNames(rep(0, length(outcomes)), outcomes)
re_samples_nofail <- MASS:mvrnorm(1e3,mu = mu, Sigma = model_nofail$vcov)
re_samples_fail <- MASS:mvrnorm(1e3,mu = mu, Sigma = model_fail$vcov)

# Get the priors
prior <- get_priors(data = df,
time_failure = "time_failure",
failure = "failure",
horizon = 5,
interval = 1/12)

# get the outcomes types
outcome_types <- get_outcome_type(data = df, outcomes)

# Select a single subject
df_pred <- df \%>\% filter(id == 10)

mmm_predictions(
  data=df,
  outcomes=outcome_types,
  fixed_formula_nofail="~ y1 + y2 + y3 + y4 + time + sex",
  random_formula_nofail="~ 1| id",
  random_effects_nofail=re_samples_nofail,
  parameters_nofail=model_nofail$estimates,
  fixed_formula_fail="~ y1 + y2 + y3 + y4 + time + sex + time_failure",
  random_formula_fail="~ 1 | id",
  random_effects_fail=re_samples_fail,
  parameters_fail=model_fail$estimates,
  time="time",
  failure="failure",
  failure_time="time_failure",
  prior=prior,
  id="id",
  landmark=1,
  horizon=5,
  interval=1/4
)
}
}
